+++ Graphics/Geometry/CombiningGeometriesExample.xaml
@@ -7,7 +7,7 @@
 <Page 
   xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" 
   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" 
-  x:Class="Geometery.CombiningGeometriesExample" 
+  x:Class="Geometry.CombiningGeometriesExample" 
   Title="Combining Geometries">
   <Page.Resources>
     <Style TargetType="{x:Type Border}">
