+++ Graphics/Geometry/GeometryAttributeSyntaxExample.xaml
@@ -1,6 +1,6 @@
 <Page xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" 
   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-  x:Class="Geometery.GeometryAttributeSyntaxExample"
+  x:Class="Geometry.GeometryAttributeSyntaxExample"
   WindowTitle="Geometry Attribute Syntax Examples">
 
   <Page.Resources>
