+++ Graphics/Geometry/GeometryUsageExample.xaml
@@ -1,7 +1,7 @@
 <Page 
   xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" 
   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" 
-  x:Class="Geometery.GeometryUsageExample" 
+  x:Class="Geometry.GeometryUsageExample" 
   WindowTitle="Geometries"
   >
   <Page.Resources>
