+++ Graphics/Geometry/Properties/Resources.resx
