+++ Graphics/Geometry/SampleViewer.xaml
@@ -2,7 +2,7 @@
   xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" 
   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" 
   WindowTitle="Geometries Sample"
-  x:Class="Geometery.SampleViewer"
+  x:Class="Geometry.SampleViewer"
   Loaded="PageLoaded">
 
   <Page.Resources>
