+++ Graphics/Geometry/ShapeGeometriesExample.xaml
@@ -1,7 +1,7 @@
 <Page 
   xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" 
   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-  x:Class="Geometery.ShapeGeometriesExample"
+  x:Class="Geometry.ShapeGeometriesExample"
     Title="Shape Geometries"
   >
   <Page.Resources>
